Crate libxml

A wrapper for libxml2

This library provides an interface to a subset of the libxml API. The idea is to extend it whenever more functionality is needed. Providing a more or less complete wrapper would be too much work.

Modules

global
Global memory management

parser
XML and HTML parsing The parser functionality

tree
Manipulations on the DOM representation The tree functionality

xpath
XPath module for global lookup in the DOM The XPath functionality