Struct libxml::bindings::_xmlXPathContext

#[repr(C)]pub struct _xmlXPathContext {
    pub doc: xmlDocPtr,
    pub node: xmlNodePtr,
    pub nb_variables_unused: c_int,
    pub max_variables_unused: c_int,
    pub varHash: xmlHashTablePtr,
    pub nb_types: c_int,
    pub max_types: c_int,
    pub types: xmlXPathTypePtr,
    pub nb_funcs_unused: c_int,
    pub max_funcs_unused: c_int,
    pub funcHash: xmlHashTablePtr,
    pub nb_axis: c_int,
    pub max_axis: c_int,
    pub axis: xmlXPathAxisPtr,
    pub namespaces: *mut xmlNsPtr,
    pub nsNr: c_int,
    pub user: *mut c_void,
    pub contextSize: c_int,
    pub proximityPosition: c_int,
    pub xptr: c_int,
    pub here: xmlNodePtr,
    pub origin: xmlNodePtr,
    pub nsHash: xmlHashTablePtr,
    pub varLookupFunc: xmlXPathVariableLookupFunc,
    pub varLookupData: *mut c_void,
    pub extra: *mut c_void,
    pub function: *const xmlChar,
    pub functionURI: *const xmlChar,
    pub funcLookupFunc: xmlXPathFuncLookupFunc,
    pub funcLookupData: *mut c_void,
    pub tmpNsList: *mut xmlNsPtr,
    pub tmpNsNr: c_int,
    pub userData: *mut c_void,
    pub error: xmlStructuredErrorFunc,
    pub lastError: xmlError,
    pub debugNode: xmlNodePtr,
    pub dict: xmlDictPtr,
    pub flags: c_int,
    pub cache: *mut c_void,
}

xmlXPathContext:

Expression evaluation occurs with respect to a context. he context consists of:

a node (the context node)

a node list (the context node list)

a set of variable bindings

a function library

the set of namespace declarations in scope for the expression Following the switch to hash tables, this need to be trimmed up at the next binary incompatible release. The node may be modified when the context is passed to libxml2 for an XPath evaluation so you may need to initialize it again before the next call.

Fields
doc: xmlDocPtrnode: xmlNodePtrnb_variables_unused: c_intmax_variables_unused: c_intvarHash: xmlHashTablePtrnb_types: c_intmax_types: c_inttypes: xmlXPathTypePtrnb_funcs_unused: c_intmax_funcs_unused: c_intfuncHash: xmlHashTablePtrnb_axis: c_intmax_axis: c_intaxis: xmlXPathAxisPtrnamespaces: *mut xmlNsPtrnsNr: c_intuser: *mut c_voidcontextSize: c_intproximityPosition: c_intxptr: c_inthere: xmlNodePtrorigin: xmlNodePtrnsHash: xmlHashTablePtrvarLookupFunc: xmlXPathVariableLookupFuncvarLookupData: *mut c_voidextra: *mut c_voidfunction: *const xmlCharfunctionURI: *const xmlCharfuncLookupFunc: xmlXPathFuncLookupFuncfuncLookupData: *mut c_voidtmpNsList: *mut xmlNsPtrtmpNsNr: c_intuserData: *mut c_voiderror: xmlStructuredErrorFunclastError: xmlErrordebugNode: xmlNodePtrdict: xmlDictPtrflags: c_intcache: *mut c_void