libxml::bindings

Type Alias docbParserCtxt

pub type docbParserCtxt = xmlParserCtxt;

xmlParserCtxt:

The parser context. NOTE This doesn’t completely define the parser state, the (current ?) design of the parser uses recursive function calls since this allow and easy mapping from the production rules of the specification to the actual code. The drawback is that the actual function call also reflect the parser state. However most of the parsing routines takes as the only argument the parser context pointer, so migrating to a state based parser for progressive parsing shouldn’t be too hard.

Aliased Type

struct docbParserCtxt {

    pub sax: *mut _xmlSAXHandler,
    pub userData: *mut c_void,
    pub myDoc: *mut _xmlDoc,
    pub wellFormed: i32,
    pub replaceEntities: i32,
    pub version: *const u8,
    pub encoding: *const u8,
    pub standalone: i32,
    pub html: i32,
    pub input: *mut _xmlParserInput,
    pub inputNr: i32,
    pub inputMax: i32,
    pub inputTab: *mut *mut _xmlParserInput,
    pub node: *mut _xmlNode,
    pub nodeNr: i32,
    pub nodeMax: i32,
    pub nodeTab: *mut *mut _xmlNode,
    pub record_info: i32,
    pub node_seq: _xmlParserNodeInfoSeq,
    pub errNo: i32,
    pub hasExternalSubset: i32,
    pub hasPErefs: i32,
    pub external: i32,
    pub valid: i32,
    pub validate: i32,
    pub vctxt: _xmlValidCtxt,
    pub instate: i32,
    pub token: i32,
    pub directory: *mut i8,
    pub name: *const u8,
    pub nameNr: i32,
    pub nameMax: i32,
    pub nameTab: *mut *const u8,
    pub nbChars: i64,
    pub checkIndex: i64,
    pub keepBlanks: i32,
    pub disableSAX: i32,
    pub inSubset: i32,
    pub intSubName: *const u8,
    pub extSubURI: *mut u8,
    pub extSubSystem: *mut u8,
    pub space: *mut i32,
    pub spaceNr: i32,
    pub spaceMax: i32,
    pub spaceTab: *mut i32,
    pub depth: i32,
    pub entity: *mut _xmlParserInput,
    pub charset: i32,
    pub nodelen: i32,
    pub nodemem: i32,
    pub pedantic: i32,
    pub _private: *mut c_void,
    pub loadsubset: i32,
    pub linenumbers: i32,
    pub catalogs: *mut c_void,
    pub recovery: i32,
    pub progressive: i32,
    pub dict: *mut _xmlDict,
    pub atts: *mut *const u8,
    pub maxatts: i32,
    pub docdict: i32,
    pub str_xml: *const u8,
    pub str_xmlns: *const u8,
    pub str_xml_ns: *const u8,
    pub sax2: i32,
    pub nsNr: i32,
    pub nsMax: i32,
    pub nsTab: *mut *const u8,
    pub attallocs: *mut i32,
    pub pushTab: *mut _xmlStartTag,
    pub attsDefault: *mut _xmlHashTable,
    pub attsSpecial: *mut _xmlHashTable,
    pub nsWellFormed: i32,
    pub options: i32,
    pub dictNames: i32,
    pub freeElemsNr: i32,
    pub freeElems: *mut _xmlNode,
    pub freeAttrsNr: i32,
    pub freeAttrs: *mut _xmlAttr,
    pub lastError: _xmlError,
    pub parseMode: u32,
    pub nbentities: u64,
    pub sizeentities: u64,
    pub nodeInfo: *mut _xmlParserNodeInfo,
    pub nodeInfoNr: i32,
    pub nodeInfoMax: i32,
    pub nodeInfoTab: *mut _xmlParserNodeInfo,
    pub input_id: i32,
    pub sizeentcopy: u64,

}

Fields
sax: *mut _xmlSAXHandleruserData: *mut c_voidmyDoc: *mut _xmlDocwellFormed: i32replaceEntities: i32version: *const u8encoding: *const u8standalone: i32html: i32input: *mut _xmlParserInputinputNr: i32inputMax: i32inputTab: *mut *mut _xmlParserInputnode: *mut _xmlNodenodeNr: i32nodeMax: i32nodeTab: *mut *mut _xmlNoderecord_info: i32node_seq: _xmlParserNodeInfoSeqerrNo: i32hasExternalSubset: i32hasPErefs: i32external: i32valid: i32validate: i32vctxt: _xmlValidCtxtinstate: i32token: i32directory: *mut i8name: *const u8nameNr: i32nameMax: i32nameTab: *mut *const u8nbChars: i64checkIndex: i64keepBlanks: i32disableSAX: i32inSubset: i32intSubName: *const u8extSubURI: *mut u8extSubSystem: *mut u8space: *mut i32spaceNr: i32spaceMax: i32spaceTab: *mut i32depth: i32entity: *mut _xmlParserInputcharset: i32nodelen: i32nodemem: i32pedantic: i32_private: *mut c_voidloadsubset: i32linenumbers: i32catalogs: *mut c_voidrecovery: i32progressive: i32dict: *mut _xmlDictatts: *mut *const u8maxatts: i32docdict: i32str_xml: *const u8str_xmlns: *const u8str_xml_ns: *const u8sax2: i32nsNr: i32nsMax: i32nsTab: *mut *const u8attallocs: *mut i32pushTab: *mut _xmlStartTagattsDefault: *mut _xmlHashTableattsSpecial: *mut _xmlHashTablensWellFormed: i32options: i32dictNames: i32freeElemsNr: i32freeElems: *mut _xmlNodefreeAttrsNr: i32freeAttrs: *mut _xmlAttrlastError: _xmlErrorparseMode: u32nbentities: u64sizeentities: u64nodeInfo: *mut _xmlParserNodeInfonodeInfoNr: i32nodeInfoMax: i32nodeInfoTab: *mut _xmlParserNodeInfoinput_id: i32sizeentcopy: u64