libxml::bindings

Type Alias xlinkHandler

pub type xlinkHandler = _xlinkHandler;

This is the structure containing a set of Links detection callbacks.

There is no default xlink callbacks, if one want to get link recognition activated, those call backs must be provided before parsing.

Aliased Type

struct xlinkHandler {
    pub simple: Option<unsafe extern "C" fn(_: *mut c_void, _: *mut _xmlNode, _: *mut u8, _: *mut u8, _: *mut u8)>,
    pub extended: Option<unsafe extern "C" fn(_: *mut c_void, _: *mut _xmlNode, _: i32, _: *const *mut u8, _: *const *mut u8, _: i32, _: *const *mut u8, _: *const *mut u8, _: *mut u32, _: *mut u32, _: i32, _: *const *mut u8, _: *mut *const u8)>,
    pub set: Option<unsafe extern "C" fn(_: *mut c_void, _: *mut _xmlNode, _: i32, _: *const *mut u8, _: *const *mut u8, _: i32, _: *const *mut u8, _: *mut *const u8)>,
}

Fields
simple: Option<unsafe extern "C" fn(_: *mut c_void, _: *mut _xmlNode, _: *mut u8, _: *mut u8, _: *mut u8)>extended: Option<unsafe extern "C" fn(_: *mut c_void, _: *mut _xmlNode, _: i32, _: *const *mut u8, _: *const *mut u8, _: i32, _: *const *mut u8, _: *const *mut u8, _: *mut u32, _: *mut u32, _: i32, _: *const *mut u8, _: *mut *const u8)>set: Option<unsafe extern "C" fn(_: *mut c_void, _: *mut _xmlNode, _: i32, _: *const *mut u8, _: *const *mut u8, _: i32, _: *const *mut u8, _: *mut *const u8)>