libxml::bindings

Type Alias xmlURI

pub type xmlURI = _xmlURI;

xmlURI:

A parsed URI reference. This is a struct containing the various fields as described in RFC 2396 but separated for further processing.

Note: query is a deprecated field which is incorrectly unescaped. query_raw takes precedence over query if the former is set. See: http://mail.gnome.org/archives/xml/2007-April/thread.html#00127

Aliased Type

struct xmlURI {
    pub scheme: *mut i8,
    pub opaque: *mut i8,
    pub authority: *mut i8,
    pub server: *mut i8,
    pub user: *mut i8,
    pub port: i32,
    pub path: *mut i8,
    pub query: *mut i8,
    pub fragment: *mut i8,
    pub cleanup: i32,
    pub query_raw: *mut i8,
}

Fields
scheme: *mut i8opaque: *mut i8authority: *mut i8server: *mut i8user: *mut i8port: i32path: *mut i8query: *mut i8fragment: *mut i8cleanup: i32query_raw: *mut i8