libxml::bindings

Type Alias xmlXPathContext

pub type xmlXPathContext = _xmlXPathContext;

xmlXPathContext:

Expression evaluation occurs with respect to a context. he context consists of:

a node (the context node)

a node list (the context node list)

a set of variable bindings

a function library

the set of namespace declarations in scope for the expression Following the switch to hash tables, this need to be trimmed up at the next binary incompatible release. The node may be modified when the context is passed to libxml2 for an XPath evaluation so you may need to initialize it again before the next call.

Aliased Type

struct xmlXPathContext {

    pub doc: *mut _xmlDoc,
    pub node: *mut _xmlNode,
    pub nb_variables_unused: i32,
    pub max_variables_unused: i32,
    pub varHash: *mut _xmlHashTable,
    pub nb_types: i32,
    pub max_types: i32,
    pub types: *mut _xmlXPathType,
    pub nb_funcs_unused: i32,
    pub max_funcs_unused: i32,
    pub funcHash: *mut _xmlHashTable,
    pub nb_axis: i32,
    pub max_axis: i32,
    pub axis: *mut _xmlXPathAxis,
    pub namespaces: *mut *mut _xmlNs,
    pub nsNr: i32,
    pub user: *mut c_void,
    pub contextSize: i32,
    pub proximityPosition: i32,
    pub xptr: i32,
    pub here: *mut _xmlNode,
    pub origin: *mut _xmlNode,
    pub nsHash: *mut _xmlHashTable,
    pub varLookupFunc: Option<unsafe extern "C" fn(_: *mut c_void, _: *const u8, _: *const u8) -> *mut _xmlXPathObject>,
    pub varLookupData: *mut c_void,
    pub extra: *mut c_void,
    pub function: *const u8,
    pub functionURI: *const u8,
    pub funcLookupFunc: Option<unsafe extern "C" fn(_: *mut c_void, _: *const u8, _: *const u8) -> Option<unsafe extern "C" fn(_: *mut _xmlXPathParserContext, _: i32)>>,
    pub funcLookupData: *mut c_void,
    pub tmpNsList: *mut *mut _xmlNs,
    pub tmpNsNr: i32,
    pub userData: *mut c_void,
    pub error: Option<unsafe extern "C" fn(_: *mut c_void, _: *mut _xmlError)>,
    pub lastError: _xmlError,
    pub debugNode: *mut _xmlNode,
    pub dict: *mut _xmlDict,
    pub flags: i32,
    pub cache: *mut c_void,
    pub opLimit: u64,
    pub opCount: u64,
    pub depth: i32,

}

Fields
doc: *mut _xmlDocnode: *mut _xmlNodenb_variables_unused: i32max_variables_unused: i32varHash: *mut _xmlHashTablenb_types: i32max_types: i32types: *mut _xmlXPathTypenb_funcs_unused: i32max_funcs_unused: i32funcHash: *mut _xmlHashTablenb_axis: i32max_axis: i32axis: *mut _xmlXPathAxisnamespaces: *mut *mut _xmlNsnsNr: i32user: *mut c_voidcontextSize: i32proximityPosition: i32xptr: i32here: *mut _xmlNodeorigin: *mut _xmlNodensHash: *mut _xmlHashTablevarLookupFunc: Option<unsafe extern "C" fn(_: *mut c_void, _: *const u8, _: *const u8) -> *mut _xmlXPathObject>varLookupData: *mut c_voidextra: *mut c_voidfunction: *const u8functionURI: *const u8funcLookupFunc: Option<unsafe extern "C" fn(_: *mut c_void, _: *const u8, _: *const u8) -> Option<unsafe extern "C" fn(_: *mut _xmlXPathParserContext, _: i32)>>funcLookupData: *mut c_voidtmpNsList: *mut *mut _xmlNstmpNsNr: i32userData: *mut c_voiderror: Option<unsafe extern "C" fn(_: *mut c_void, _: *mut _xmlError)>lastError: _xmlErrordebugNode: *mut _xmlNodedict: *mut _xmlDictflags: i32cache: *mut c_voidopLimit: u64opCount: u64depth: i32