libxml::readonly

Struct RoNode

pub struct RoNode(/* private fields */);

Lightweight struct for read-only parallel processing

Implementations

impl RoNode

pub fn node_ptr(&self) -> xmlNodePtr

Immutably borrows the underlying libxml2 xmlNodePtr pointer

pub fn get_next_sibling(self) -> Option<RoNode>

Returns the next sibling if it exists

pub fn get_prev_sibling(self) -> Option<RoNode>

Returns the previous sibling if it exists

pub fn get_first_child(self) -> Option<RoNode>

Returns the first child if it exists

pub fn get_last_child(self) -> Option<RoNode>

Returns the last child if it exists

pub fn get_next_element_sibling(&self) -> Option<RoNode>

Returns the next element sibling if it exists

pub fn get_prev_element_sibling(&self) -> Option<RoNode>

Returns the previous element sibling if it exists

pub fn get_first_element_child(self) -> Option<RoNode>

Returns the first element child if it exists

pub fn get_last_element_child(&self) -> Option<RoNode>

Returns the last element child if it exists

pub fn get_child_nodes(self) -> Vec<RoNode>

Returns all child nodes of the given node as a vector

pub fn get_child_elements(self) -> Vec<RoNode>

Returns all child elements of the given node as a vector

pub fn get_parent(self) -> Option<RoNode>

Returns the parent if it exists

pub fn get_type(self) -> Option<NodeType>

Get the node type

pub fn is_text_node(self) -> bool

Returns true if it is a text node

pub fn is_element_node(self) -> bool

Checks if the given node is an Element

pub fn is_null(self) -> bool

Checks if the underlying libxml2 pointer is NULL

pub fn get_name(self) -> String

Returns the name of the node (empty string if name pointer is NULL)

pub fn get_content(self) -> String

Returns the content of the node (assumes UTF-8 XML document)

pub fn get_property(self, name: &str) -> Option<String>

Returns the value of property name

pub fn get_property_ns(self, name: &str, ns: &str) -> Option<String>

Returns the value of property name in namespace ns

pub fn get_property_no_ns(self, name: &str) -> Option<String>

Returns the value of property name with no namespace

pub fn get_property_node(self, name: &str) -> Option<RoNode>

Return an attribute as a Node struct of type AttributeNode

pub fn get_property_node_ns(self, name: &str, ns: &str) -> Option<RoNode>

Return an attribute in a namespace ns as a Node of type AttributeNode

pub fn get_property_node_no_ns(self, name: &str) -> Option<RoNode>

Return an attribute with no namespace as a Node of type AttributeNode

pub fn get_attribute(self, name: &str) -> Option<String>

Alias for get_property

pub fn get_attribute_ns(self, name: &str, ns: &str) -> Option<String>

Alias for get_property_ns

pub fn get_attribute_no_ns(self, name: &str) -> Option<String>

Alias for get_property_no_ns

pub fn get_attribute_node(self, name: &str) -> Option<RoNode>

Alias for get_property_node

pub fn get_attribute_node_ns(self, name: &str, ns: &str) -> Option<RoNode>

Alias for get_property_node_ns

pub fn get_attribute_node_no_ns(self, name: &str) -> Option<RoNode>

Alias for get_property_node_no_ns

pub fn get_properties(self) -> HashMap<String, String>

Get a copy of the attributes of this node

pub fn get_properties_ns(self) -> HashMap<(String, Option<Namespace>), String>

Get a copy of this node’s attributes and their namespaces

pub fn get_attributes(self) -> HashMap<String, String>

Alias for get_properties

pub fn get_attributes_ns(self) -> HashMap<(String, Option<Namespace>), String>

Alias for get_properties_ns

pub fn has_property(self, name: &str) -> bool

Check if a property has been defined, without allocating its value

pub fn has_property_ns(self, name: &str, ns: &str) -> bool

Check if property name in namespace ns exists

pub fn has_property_no_ns(self, name: &str) -> bool

Check if property name with no namespace exists

pub fn has_attribute(self, name: &str) -> bool

Alias for has_property

pub fn has_attribute_ns(self, name: &str, ns: &str) -> bool

Alias for has_property_ns

pub fn has_attribute_no_ns(self, name: &str) -> bool

Alias for has_property_no_ns

pub fn get_namespace(self) -> Option<Namespace>

Gets the active namespace associated of this node

pub fn get_namespaces(self, doc: &Document) -> Vec<Namespace>

Gets a list of namespaces associated with this node

pub fn get_namespace_declarations(self) -> Vec<Namespace>

Get a list of namespaces declared with this node

pub fn lookup_namespace_prefix(self, href: &str) -> Option<String>

Looks up the prefix of a namespace from its URI, basedo around a given Node

pub fn lookup_namespace_uri(self, prefix: &str) -> Option<String>

Looks up the uri of a namespace from its prefix, basedo around a given Node

pub fn get_class_names(self) -> HashSet<String>

Get a set of class names from this node’s attributes

pub fn findnodes(self, xpath: &str, owner: &Document) -> Result<Vec<RoNode>, ()>

find read-only nodes via xpath, at the specified node and a given document

pub fn is_unlinked(self) -> bool

Read-only nodes are always linked

pub fn to_hashable(self) -> usize

libc::c_void isn’t hashable and cannot be made hashable

pub fn null() -> Self

Create a mock node, used for a placeholder argument