Module libxml::tree::node

Node, and related, feature set

Structs

Node
An xml node

Statics

NODE_RC_MAX_GUARD
Guard treshold for enforcing runtime mutability checks for Nodes

Functions

set_node_rc_guard
Set the guard value for the max Rc "strong count" allowed for mutable use of a Node Default is 2