Struct libxml::tree::Node

pub struct Node {
    pub node_ptr: *mut c_void,
}

An xml node

Fields
node_ptr: *mut c_void

libxml's xmlNodePtr

Methods

impl Node

pub fn new(
name: &str,
ns: Option<Namespace>,
doc: &Document
) -> Result<Self, ()>

Create a new node, bound to a given document.

pub fn new_text(content: &str, doc: &Document) -> Result<Self, ()>

Create a new text node, bound to a given document

pub fn mock() -> Self

Create a mock node, used for a placeholder argument

pub fn to_hashable(&self) -> usize

For some reason libc::c_void isn't hashable and cannot be made hashable

pub fn get_next_sibling(&self) -> Option<Node>

Returns the next sibling if it exists

pub fn get_prev_sibling(&self) -> Option<Node>

Returns the previous sibling if it exists

pub fn get_first_child(&self) -> Option<Node>

Returns the first child if it exists

pub fn get_first_element_child(&self) -> Option<Node>

Returns the first element child if it exists

pub fn get_last_child(&self) -> Option<Node>

Returns the last child if it exists

pub fn get_child_nodes(&self) -> Vec<Node>

Returns all child nodes of the given node as a vector

pub fn get_child_elements(&self) -> Vec<Node>

Returns all child elements of the given node as a vector

pub fn get_parent(&self) -> Option<Node>

Returns the parent if it exists

pub fn get_type(&self) -> Option<NodeType>

Get the node type

pub fn add_prev_sibling(&self, new_sibling: &Node) -> Result<(), ()>

Add a previous sibling

pub fn add_next_sibling(&self, new_sibling: &Node) -> Result<(), ()>

Add a next sibling

pub fn is_text_node(&self) -> bool

Returns true iff it is a text node

pub fn is_element_node(&self) -> bool

Checks if the given node is an Element

pub fn get_name(&self) -> String

Returns the name of the node (empty string if name pointer is NULL)

pub fn set_name(&mut self, name: &str)

Sets the name of this Node

pub fn get_content(&self) -> String

Returns the content of the node (empty string if content pointer is NULL)

pub fn set_content(&mut self, content: &str)

Sets the text content of this Node

pub fn get_property(&self, name: &str) -> Option<String>

Returns the value of property name

pub fn get_property_ns(&self, name: &str, ns: &str) -> Option<String>

Returns the value of property name in namespace ns

pub fn get_property_node(&self, name: &str) -> Option<Node>

Return an attribute as a Node struct of type AttributeNode

pub fn set_property(&mut self, name: &str, value: &str)

Sets the value of property name to value

pub fn set_property_ns(&mut self, name: &str, value: &str, ns: &Namespace)

Sets a namespaced attribute

pub fn remove_property(&mut self, name: &str)

Removes the property of given name

pub fn get_attribute(&self, name: &str) -> Option<String>

Alias for get_property

pub fn get_attribute_ns(&self, name: &str, ns: &str) -> Option<String>

Alias for get_property_ns

pub fn get_attribute_node(&self, name: &str) -> Option<Node>

Alias for get_property_node

pub fn set_attribute(&mut self, name: &str, value: &str)

Alias for set_property

pub fn set_attribute_ns(&mut self, name: &str, value: &str, ns: &Namespace)

Alias for set_property_ns

pub fn remove_attribute(&mut self, name: &str)

Alias for remove_property

pub fn get_properties(&self) -> HashMap<String, String>

Get a copy of the attributes of this node

pub fn get_attributes(&self) -> HashMap<String, String>

Alias for get_properties

pub fn get_namespace(&self) -> Option<Namespace>

Gets the active namespace associated of this node

pub fn get_namespaces(&self, doc: &Document) -> Vec<Namespace>

Gets a list of namespaces associated with this node

pub fn get_namespace_declarations(&self) -> Vec<Namespace>

Get a list of namespaces declared with this node

pub fn set_namespace(&mut self, namespace: &Namespace)

Sets a Namespace for the node

pub fn lookup_namespace_prefix(&self, href: &str) -> Option<String>

Looks up the prefix of a namespace from its URI, basedo around a given Node

pub fn lookup_namespace_uri(&self, prefix: &str) -> Option<String>

Looks up the uri of a namespace from its prefix, basedo around a given Node

pub fn recursively_remove_namespaces(&mut self)

Removes the namespaces of this Node and it's children!

pub fn get_class_names(&self) -> HashSet<String>

Get a set of class names from this node's attributes

pub fn add_child(&mut self, child: Node) -> Result<Node, ()>

Creates a new Node as child to the self Node

pub fn new_child(
&mut self,
ns: Option<Namespace>,
name: &str
) -> Result<Node, ()>

Creates a new Node as child to the self Node

pub fn add_text_child(
&mut self,
ns: Option<Namespace>,
name: &str,
content: &str
) -> Result<Node, ()>

Adds a new text child, to this Node

pub fn append_text(&mut self, content: &str)

Append text to this Node

pub fn unlink_node(&mut self)

Unbinds the Node from its siblings and Parent, but not from the Document it belongs to. If the node is not inserted into the DOM afterwards, it will be lost after the program terminates. From a low level view, the unbound node is stripped from the context it is and inserted into a (hidden) document-fragment.

pub fn unlink(&mut self)

Alias for unlink_node

pub fn unbind_node(&mut self)

Alias for unlink_node

pub fn unbind(&mut self)

Alias for unlink_node