libxml

Module xpath

XPath module for global lookup in the DOM The XPath functionality

Structs

Context

An XPath context

Object

Essentially, the result of the evaluation of some xpath expression

Functions

is_well_formed_xpath

Calls the binding to http://xmlsoft.org/html/libxml-xpath.html#xmlXPathCompile and return true if a non-null pointer is returned. The idea is to use this to validate an xpath independent of context. Tests describing what this validates in tests/xpath_tests.rs