Struct libxml::xpath::Context

pub struct Context<'a> {
    pub context_ptr: *mut c_void,
    pub document: &'a Document,
}

The xpath context

Fields
context_ptr: *mut c_void

libxml's ContextPtr

document: &'a Document

Document contains pointer, needed for ContextPtr, so we need to borrow Document to prevent it's freeing

Methods

impl<'a> Context<'a>

fn new(doc: &Document) -> Result<Context, ()>

create the xpath context for a document

fn register_namespace(&self, prefix: &str, href: &str) -> Result<(), ()>

Register a namespace prefix-href pair on the xpath context

fn evaluate(&self, xpath: &str) -> Result<Object, ()>

evaluate an xpath

fn set_context_node(&mut self, node: &Node) -> Result<(), ()>

localize xpath context to a specific Node

fn findnodes(
&mut self,
xpath: &str,
node_opt: Option<&Node>
) -> Result<Vec<Node>, ()>

find nodes via xpath, at a specified node or the document root

fn findvalue(
&mut self,
xpath: &str,
node_opt: Option<&Node>
) -> Result<String, ()>

find a literal value via xpath, at a specified node or the document root