Struct libxml::xpath::Object

pub struct Object {
    pub ptr: *mut c_void,
}

Essentially, the result of the evaluation of some xpath expression

Fields
ptr: *mut c_void

libxml's ObjectPtr

Methods

impl Object

fn get_number_of_nodes(&self) -> usize

get the number of nodes in the result set

fn get_nodes_as_vec(&self) -> Vec<Node>

returns the result set as a vector of node references

fn to_string(&self) -> String

use if the XPath used was meant to return a string, such as string(//foo/@attr)