Crate libxml

A wrapper for libxml2

This library provides an interface to a subset of the libxml API. The idea is to extend it whenever more functionality is needed. Providing a more or less complete wrapper would be too much work.

Modules

bindings
Bindings to the C interface

error
XML Global Error Structures and Handling

parser
XML and HTML parsing The parser functionality

readonly
Read-only parallel primitives

schemas
Schema Validation

tree
Manipulations on the DOM representation The tree functionality

xpath
XPath module for global lookup in the DOM The XPath functionality