libxml::bindings

Struct _xmlParserCtxt

#[repr(C)]
pub struct _xmlParserCtxt {

    pub sax: *mut _xmlSAXHandler,
    pub userData: *mut c_void,
    pub myDoc: xmlDocPtr,
    pub wellFormed: c_int,
    pub replaceEntities: c_int,
    pub version: *const xmlChar,
    pub encoding: *const xmlChar,
    pub standalone: c_int,
    pub html: c_int,
    pub input: xmlParserInputPtr,
    pub inputNr: c_int,
    pub inputMax: c_int,
    pub inputTab: *mut xmlParserInputPtr,
    pub node: xmlNodePtr,
    pub nodeNr: c_int,
    pub nodeMax: c_int,
    pub nodeTab: *mut xmlNodePtr,
    pub record_info: c_int,
    pub node_seq: xmlParserNodeInfoSeq,
    pub errNo: c_int,
    pub hasExternalSubset: c_int,
    pub hasPErefs: c_int,
    pub external: c_int,
    pub valid: c_int,
    pub validate: c_int,
    pub vctxt: xmlValidCtxt,
    pub instate: xmlParserInputState,
    pub token: c_int,
    pub directory: *mut c_char,
    pub name: *const xmlChar,
    pub nameNr: c_int,
    pub nameMax: c_int,
    pub nameTab: *mut *const xmlChar,
    pub nbChars: c_long,
    pub checkIndex: c_long,
    pub keepBlanks: c_int,
    pub disableSAX: c_int,
    pub inSubset: c_int,
    pub intSubName: *const xmlChar,
    pub extSubURI: *mut xmlChar,
    pub extSubSystem: *mut xmlChar,
    pub space: *mut c_int,
    pub spaceNr: c_int,
    pub spaceMax: c_int,
    pub spaceTab: *mut c_int,
    pub depth: c_int,
    pub entity: xmlParserInputPtr,
    pub charset: c_int,
    pub nodelen: c_int,
    pub nodemem: c_int,
    pub pedantic: c_int,
    pub _private: *mut c_void,
    pub loadsubset: c_int,
    pub linenumbers: c_int,
    pub catalogs: *mut c_void,
    pub recovery: c_int,
    pub progressive: c_int,
    pub dict: xmlDictPtr,
    pub atts: *mut *const xmlChar,
    pub maxatts: c_int,
    pub docdict: c_int,
    pub str_xml: *const xmlChar,
    pub str_xmlns: *const xmlChar,
    pub str_xml_ns: *const xmlChar,
    pub sax2: c_int,
    pub nsNr: c_int,
    pub nsMax: c_int,
    pub nsTab: *mut *const xmlChar,
    pub attallocs: *mut c_int,
    pub pushTab: *mut xmlStartTag,
    pub attsDefault: xmlHashTablePtr,
    pub attsSpecial: xmlHashTablePtr,
    pub nsWellFormed: c_int,
    pub options: c_int,
    pub dictNames: c_int,
    pub freeElemsNr: c_int,
    pub freeElems: xmlNodePtr,
    pub freeAttrsNr: c_int,
    pub freeAttrs: xmlAttrPtr,
    pub lastError: xmlError,
    pub parseMode: xmlParserMode,
    pub nbentities: c_ulong,
    pub sizeentities: c_ulong,
    pub nodeInfo: *mut xmlParserNodeInfo,
    pub nodeInfoNr: c_int,
    pub nodeInfoMax: c_int,
    pub nodeInfoTab: *mut xmlParserNodeInfo,
    pub input_id: c_int,
    pub sizeentcopy: c_ulong,

}

xmlParserCtxt:

The parser context. NOTE This doesn’t completely define the parser state, the (current ?) design of the parser uses recursive function calls since this allow and easy mapping from the production rules of the specification to the actual code. The drawback is that the actual function call also reflect the parser state. However most of the parsing routines takes as the only argument the parser context pointer, so migrating to a state based parser for progressive parsing shouldn’t be too hard.

Fields
sax: *mut _xmlSAXHandleruserData: *mut c_voidmyDoc: xmlDocPtrwellFormed: c_intreplaceEntities: c_intversion: *const xmlCharencoding: *const xmlCharstandalone: c_inthtml: c_intinput: xmlParserInputPtrinputNr: c_intinputMax: c_intinputTab: *mut xmlParserInputPtrnode: xmlNodePtrnodeNr: c_intnodeMax: c_intnodeTab: *mut xmlNodePtrrecord_info: c_intnode_seq: xmlParserNodeInfoSeqerrNo: c_inthasExternalSubset: c_inthasPErefs: c_intexternal: c_intvalid: c_intvalidate: c_intvctxt: xmlValidCtxtinstate: xmlParserInputStatetoken: c_intdirectory: *mut c_charname: *const xmlCharnameNr: c_intnameMax: c_intnameTab: *mut *const xmlCharnbChars: c_longcheckIndex: c_longkeepBlanks: c_intdisableSAX: c_intinSubset: c_intintSubName: *const xmlCharextSubURI: *mut xmlCharextSubSystem: *mut xmlCharspace: *mut c_intspaceNr: c_intspaceMax: c_intspaceTab: *mut c_intdepth: c_intentity: xmlParserInputPtrcharset: c_intnodelen: c_intnodemem: c_intpedantic: c_int_private: *mut c_voidloadsubset: c_intlinenumbers: c_intcatalogs: *mut c_voidrecovery: c_intprogressive: c_intdict: xmlDictPtratts: *mut *const xmlCharmaxatts: c_intdocdict: c_intstr_xml: *const xmlCharstr_xmlns: *const xmlCharstr_xml_ns: *const xmlCharsax2: c_intnsNr: c_intnsMax: c_intnsTab: *mut *const xmlCharattallocs: *mut c_intpushTab: *mut xmlStartTagattsDefault: xmlHashTablePtrattsSpecial: xmlHashTablePtrnsWellFormed: c_intoptions: c_intdictNames: c_intfreeElemsNr: c_intfreeElems: xmlNodePtrfreeAttrsNr: c_intfreeAttrs: xmlAttrPtrlastError: xmlErrorparseMode: xmlParserModenbentities: c_ulongsizeentities: c_ulongnodeInfo: *mut xmlParserNodeInfonodeInfoNr: c_intnodeInfoMax: c_intnodeInfoTab: *mut xmlParserNodeInfoinput_id: c_intsizeentcopy: c_ulong